engine2 msg split
